tidying, login prompt for liked songs
