+++ web/src/app.js
@@ -2,9 +2,7 @@ import React from "react";
 import { Routes, Route } from "react-router-dom";
 import Start from "containers/start";
 import Login from "containers/login";
-import Host from "containers/host";
 import { Room } from "containers/room";
-import Join from "containers/join";
 import { RecoilRoot } from "recoil";
 import { PersistenceObserver, initializeState } from "./store";
 
@@ -15,8 +13,6 @@ function App() {
       <Routes>
         <Route path="/" element={<Start />} />
         <Route path="login" element={<Login />} />
-        <Route path="host" element={<Host />} />
-        <Route path="join/:id" element={<Join />} />
         <Route path="rooms/:id/*" element={<Room />} />
       </Routes>
     </RecoilRoot>
