+++ web/src/hooks/spotify/use-liked-songs.js
@@ -1,20 +1,24 @@
 import { queryCache, useMutation, useQuery } from "react-query";
 import { useRecoilValue } from "recoil";
-import { userAccessTokenState } from "hooks/use-login";
+import { userAccessTokenState } from "hooks/spotify/use-login";
 import axios from "axios";
 
 const { REACT_APP_SPOTIFY_API_URL: SPOTIFY_API_URL } = process.env;
 
-export function useLikedSongs() {
+export function useGetLikedSongs() {
   const userAccessToken = useRecoilValue(userAccessTokenState);
-  return useQuery(userAccessToken && "likes", async () => {
-    const { data } = await axios.get(`${SPOTIFY_API_URL}/me/tracks`, {
-      headers: {
-        Authorization: `Bearer ${userAccessToken}`,
-      },
-    });
-    return data.items;
-  });
+  return useQuery(
+    userAccessToken && "likes",
+    async () => {
+      const { data } = await axios.get(`${SPOTIFY_API_URL}/me/tracks`, {
+        headers: {
+          Authorization: `Bearer ${userAccessToken}`,
+        },
+      });
+      return data.items;
+    },
+    { retry: false, refetchOnWindowFocus: false },
+  );
 }
 
 export function useCheckIfSavedTrack(trackId) {
