--- web/src/hooks/use-playlists.js
@@ -1,21 +1,0 @@
-import { useQuery } from "react-query";
-import { useRecoilValue } from "recoil";
-import { userAccessTokenState } from "hooks/use-login";
-import axios from "axios";
-
-const { REACT_APP_SPOTIFY_API_URL: SPOTIFY_API_URL } = process.env;
-
-export function usePlaylists(userDetails) {
-  const userAccessToken = useRecoilValue(userAccessTokenState);
-  return useQuery(userDetails && ["playlists", userDetails.id], async () => {
-    const { data } = await axios.get(
-      `${SPOTIFY_API_URL}/users/${userDetails.id}/playlists`,
-      {
-        headers: {
-          Authorization: `Bearer ${userAccessToken}`,
-        },
-      },
-    );
-    return data;
-  });
-}
